+++ user_docs/en/changes.md
@@ -85,6 +85,7 @@ Add-ons will need to be re-tested and have their manifest updated.
 On ARM64 machines with Windows 11, these ARM64EC libraries are loaded instead of their X64 equivalents. (#18570, @leonarddeR)
 * NVDA is now licensed under "GPL-2 or later".
 * In `braille.py`, the `FormattingMarker` class has a new `shouldBeUsed` method, to determine if the formatting marker key should be reported (#7608, @nvdaes)
+* Added `api.fakeNVDAObject` set and `api.isFakeNVDAObject` function to identify fake NVDAObject instances. (@hwf1324)
 
 #### API Breaking Changes